JobScraper/pyproject.toml
[project]
name = "jobscraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "dspy-ai>=3.1.0",
    "faiss-cpu>=1.13.2",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "langchain>=1.2.6",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.1",
    "langchain-openai>=1.1.7",
    "langgraph>=1.0.6",
    "lxml>=6.0.2",
    "playwright>=1.57.0",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.0",
    "types-beautifulsoup4>=4.12.0.20250516",
    "types-requests>=2.32.4.20260107",
    "uvicorn>=0.40.0",
]
